move back to single main, no cmd
