+++ Dockerfile
@@ -24,7 +24,7 @@ RUN go mod tidy; go mod download
 # -o app: Specifies the output executable name as 'app'.
 # -ldflags "-s -w": Reduces the size of the binary by stripping debug information and symbol tables.
 # CGO_ENABLED=0: Disables CGO, making the binary statically linked and truly independent of the base image.
-RUN FLUFFY_TEMPLATE_DIR=cmd/gui/ CGO_ENABLED=0 GOOS=linux go build -o app -ldflags "-s -w" cmd/gui/*.go
+RUN CGO_ENABLED=0 GOOS=linux go build -o app -ldflags "-s -w" main.go
 
 # Create the final minimal image
 # We use 'scratch' which is an empty base image, smaller footprint
@@ -35,8 +35,8 @@ WORKDIR /root/
 
 # Copy the compiled executable from the 'builder' stage
 COPY --from=builder /go/src/github.com/papaburgs/fluffy-robot/app .
-COPY --from=builder /go/src/github.com/papaburgs/fluffy-robot/cmd/gui/templates ./templates
-# COPY --from=builder /go/src/github.com/papaburgs/fluffy-robot/static ./
+# COPY --from=builder /go/src/github.com/papaburgs/fluffy-robot/templates ./templates/
+# COPY --from=builder /go/src/github.com/papaburgs/fluffy-robot/static ./static/
 # Copy the CA certificates from the builder stage
 # Standard location for many applications to find them
 COPY --from=builder /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
