--- cmd/gui/static/style.css
@@ -1,393 +1,0 @@
-:root {
-    --bg-color: #121212;
-    --text-color: #E0E0E0;
-    --primary-color: #1F3C88;
-    --accent-color: #bb86fc;
-    --sidebar-width: 250px;
-    --card-bg: #1E1E1E;
-    --hover-bg: #333;
-}
-
-body {
-    background-color: var(--bg-color);
-    color: var(--text-color);
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-}
-
-/* App Container */
-.app-container {
-    display: flex;
-    flex: 1;
-    position: relative;
-    overflow-x: hidden;
-}
-
-/* Sidebar */
-.sidebar {
-    width: var(--sidebar-width);
-    background-color: var(--card-bg);
-    height: 100vh;
-    position: fixed;
-    top: 0;
-    left: calc(-1 * var(--sidebar-width)); /* Hidden by default on mobile */
-    transition: left 0.3s ease-in-out, margin-left 0.3s ease-in-out, width 0.3s ease-in-out;
-    z-index: 1000;
-    box-shadow: 2px 0 5px rgba(0,0,0,0.5);
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-}
-
-.sidebar.open {
-    left: 0 !important;
-}
-
-.sidebar-header {
-    padding: 20px;
-    background-color: var(--primary-color);
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.sidebar-header h3 {
-    margin: 0;
-    font-size: 1.2rem;
-}
-
-.nav-links {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    overflow-y: auto;
-}
-
-.nav-section {
-    padding: 20px 20px 5px;
-    font-size: 0.75rem;
-    text-transform: uppercase;
-    color: #888;
-    letter-spacing: 1px;
-    font-weight: bold;
-}
-
-.nav-link {
-    display: block;
-    padding: 12px 20px;
-    color: var(--text-color);
-    text-decoration: none;
-    transition: background 0.2s, color 0.2s;
-    cursor: pointer;
-}
-
-.nav-link:hover {
-    background-color: var(--hover-bg);
-    color: var(--accent-color);
-}
-
-.icon {
-    margin-right: 10px;
-    font-size: 1.1rem;
-}
-
-.nav-spacer {
-    flex: 1;
-}
-
-.preferences-link {
-    border-top: 1px solid #333;
-}
-
-.icon-button {
-    background: none;
-    border: none;
-    color: var(--text-color);
-    font-size: 1.5rem;
-    cursor: pointer;
-    padding: 5px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-/* Main Content */
-main {
-    flex: 1;
-    padding: 20px;
-    width: 100%;
-    box-sizing: border-box;
-    transition: all 0.3s ease-in-out;
-}
-
-/* Header */
-header {
-    margin-bottom: 20px;
-    padding-bottom: 10px;
-    border-bottom: 1px solid #333;
-}
-
-.header-container {
-    display: flex;
-    align-items: center;
-    gap: 15px;
-}
-
-.header-content {
-    flex: 1;
-}
-
-.header-content h1 {
-    margin: 0;
-    font-size: 1.5rem;
-}
-
-.header-content p {
-    margin: 5px 0 0 0;
-    font-size: 0.9rem;
-}
-
-/* Hamburger Button */
-.hamburger {
-    background: var(--primary-color);
-    border: none;
-    color: white;
-    padding: 8px 12px;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 1.2rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transition: background 0.2s;
-}
-
-.hamburger:hover {
-    background-color: var(--hover-bg);
-}
-
-/* Desktop Sidebar behavior */
-@media (min-width: 769px) {
-    .sidebar {
-        position: relative; /* Take up space in flex container */
-        left: 0 !important;
-        margin-left: 0;
-        z-index: 10;
-        box-shadow: none;
-        border-right: 1px solid #333;
-        min-width: 0;
-    }
-
-    .app-container.sidebar-collapsed .sidebar {
-        margin-left: calc(-1 * var(--sidebar-width));
-        width: 0;
-        border-right: none;
-    }
-
-    /* Hide the 'X' button on desktop, since toggle button is available */
-    #close-sidebar {
-        display: none;
-    }
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-    main {
-        padding: 15px;
-    }
-}
-
-.chart-scroll-wrapper {
-    width: 100%;
-    overflow-x: auto;
-    -webkit-overflow-scrolling: touch; /* Smooth scrolling on iOS */
-    margin-bottom: 20px;
-    background-color: var(--card-bg);
-    border-radius: 8px;
-    padding: 10px;
-    box-sizing: border-box;
-}
-
-.chart-scroll-wrapper > div {
-    min-width: 800px; /* Force a readable width on mobile to enable scrolling */
-}
-
-/* Ensure the chart inside doesn't get squashed and allows scrolling */
-.chart-scroll-wrapper > div > div {
-    margin: 0 auto;
-}
-
-/* Construction Summary Table */
-.construction-summary {
-    margin: 30px 0;
-    padding: 20px;
-    background-color: var(--card-bg);
-    border-radius: 8px;
-    box-shadow: 0 2px 10px rgba(0,0,0,0.3);
-}
-
-.construction-summary h3 {
-    margin-top: 0;
-    color: var(--accent-color);
-}
-
-.table-container {
-    overflow-x: auto;
-}
-
-table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-top: 10px;
-}
-
-th, td {
-    padding: 12px 15px;
-    text-align: left;
-    border-bottom: 1px solid #333;
-}
-
-th {
-    background-color: rgba(31, 60, 136, 0.3);
-    color: var(--accent-color);
-    font-weight: 600;
-}
-
-tr:hover {
-    background-color: rgba(255, 255, 255, 0.05);
-}
-
-/* Responsive adjustments */
-.my-agent-highlight {
-    background-color: rgba(187, 134, 252, 0.2) !important;
-    font-weight: bold;
-}
-
-.my-agent-highlight td {
-    color: var(--accent-color) !important;
-}
-
-/* Agent Buttons in Preferences */
-.agent-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
-    gap: 10px;
-    margin-top: 20px;
-}
-
-.agent-button {
-    padding: 10px 15px;
-    border-radius: 8px;
-    border: 2px solid transparent;
-    cursor: pointer;
-    font-weight: 600;
-    transition: all 0.2s ease;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    position: relative;
-    background-color: #2a2a2a;
-    color: #888;
-}
-
-/* State: Active (credits != 175k) */
-.agent-button.active {
-    background-color: #2d3a2d;
-    color: #a0c0a0;
-    border-color: #3e4e3e;
-}
-
-/* State: Inactive (credits == 175k) */
-.agent-button.inactive {
-    background-color: #2a2a2a;
-    color: #666;
-    border-color: #333;
-    font-style: italic;
-}
-
-/* State: Selected (Added to active or inactive) */
-.agent-button.agent-selected {
-    border-color: var(--accent-color);
-    box-shadow: 0 0 8px rgba(187, 134, 252, 0.3);
-}
-
-/* Pair: Active + Selected */
-.agent-button.active.agent-selected {
-    background-color: #3d5a3d;
-    color: #fff;
-}
-
-/* Pair: Inactive + Selected */
-.agent-button.inactive.agent-selected {
-    background-color: #333;
-    color: #ddd;
-}
-
-/* Hover effects */
-.agent-button:hover {
-    transform: translateY(-2px);
-    filter: brightness(1.1);
-}
-
-.status-indicator {
-    width: 8px;
-    height: 8px;
-    border-radius: 50%;
-    margin-left: 8px;
-    background-color: #f44336; /* Default: Red (not selected) */
-    box-shadow: 0 0 5px rgba(244, 67, 54, 0.4);
-}
-
-.agent-button.agent-selected .status-indicator {
-    background-color: #4caf50; /* Selected: Green */
-    box-shadow: 0 0 5px #4caf50;
-}
-
-/* Search and Filters */
-.controls-container {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 20px;
-    align-items: center;
-    margin-bottom: 20px;
-}
-
-.search-box {
-    flex: 1;
-    min-width: 200px;
-}
-
-.search-container {
-    margin-bottom: 20px;
-}
-
-.search-input {
-    width: 100%;
-    padding: 10px 15px;
-    border-radius: 8px;
-    border: 1px solid #333;
-    background: #1e1e1e;
-    color: #fff;
-}
-
-.filter-options {
-    display: flex;
-    gap: 20px;
-    margin-bottom: 20px;
-    align-items: center;
-}
-
-.checkbox-label, .select-label {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    cursor: pointer;
-    font-size: 0.9rem;
-}
