--- cmd/gui/templates/header.html
@@ -1,9 +1,0 @@
-<header>
-    <div class="header-container">
-        <button id="toggle-sidebar" type="button" class="icon-button hamburger" aria-label="Toggle Menu">&#9776;</button>
-        <div class="header-content">
-            <h1>Fluffy Robot</h1>
-            <p>Reset: <span class="accent">{{.Reset}}</span></p>
-        </div>
-    </div>
-</header>
