--- cmd/gui/templates/jumpgates.html
@@ -1,23 +1,0 @@
-<div class="jumpgates-container">
-    <h2>Jumpgate Status</h2>
-    <table class="data-table">
-        <thead>
-            <tr>
-                <th>System</th>
-                <th>Headquarters</th>
-                <th>Jumpgate</th>
-                <th>Status</th>
-            </tr>
-        </thead>
-        <tbody>
-            {{range .}}
-            <tr class="{{if .complete}}complete{{end}}">
-                <td>{{.system}}</td>
-                <td>{{.headquarters}}</td>
-                <td>{{.jumpgate}}</td>
-                <td>{{.status}}{{if .complete}} ({{.completeTime.Format "2006-01-02 15:04"}}){{end}}</td>
-            </tr>
-            {{end}}
-        </tbody>
-    </table>
-</div>
