--- cmd/gui/templates/permissions-grid.html
@@ -1,11 +1,0 @@
-{{range .}}
-    <button id="btn-{{.Name}}" 
-            type="button"
-            class="agent-button {{if .IsActive}}active{{else}}inactive{{end}} {{if .IsChecked}}agent-selected{{end}}"
-            onclick="toggleAgentSelection('{{.Name}}')"
-            title="{{if .IsActive}}Active Agent{{else}}Inactive (175k credits){{end}}"
-    >
-        {{.Name}}
-        <span class="status-indicator"></span>
-    </button>
-{{end}}
