--- cmd/gui/templates/permissions.html
@@ -1,74 +1,0 @@
-<div class="preferences-container" hx-vals='js:{storageAgents: getStoredAgentsForHxVals(), myAgent: getMyAgentForHxVals()}' hx-trigger="refresh-vals from:body">
-    <h2>Permissions</h2>
-    
-    <h2>Dashboard Agents</h2>
-    <p>Select agents to include in your dashboard. Changes are saved automatically.</p>
-    
-    <div class="controls-container">
-        <div class="search-box">
-            <input class="search-input" type="search" 
-                   name="agentSearch" 
-                   placeholder="Search agents..." 
-                   hx-get="/permissions-grid" 
-                   hx-trigger="input changed delay:500ms, keyup[key=='Enter'], load, refresh from:#agent-grid" 
-                   hx-target="#agent-grid" 
-                   hx-indicator=".loading-indicator"
-                   hx-include="[name='hideInactive'], [name='sortBy']">
-        </div>
-        
-        <div class="filter-options">
-            <label class="checkbox-label">
-                <input type="checkbox" name="hideInactive" 
-                       hx-get="/permissions-grid" 
-                       hx-target="#agent-grid" 
-                       hx-include="[name='agentSearch'], [name='sortBy']">
-                Hide Inactive
-            </label>
-            
-            <label class="select-label">
-                Sort by:
-                <select name="sortBy" 
-                        hx-get="/permissions-grid" 
-                        hx-target="#agent-grid" 
-                        hx-include="[name='agentSearch'], [name='hideInactive']">
-                    <option value="name">Name</option>
-                    <option value="credits">Credits</option>
-                </select>
-            </label>
-        </div>
-        <span class="loading-indicator">Updating...</span>
-    </div>
-
-    <div id="agent-grid" class="agent-grid"
-         hx-get="/permissions-grid"
-         hx-trigger="load delay:10ms">
-        <!-- Agents will be loaded here via HTMX -->
-    </div>
-    </div>
-    <hr class="pref-divider">
-    <div class="main-agent-section">
-        <h3>Main Agent (Highlight)</h3>
-        <p>Select an agent to highlight in the leaderboard and dashboard.</p>
-        <div class="controls-container">
-            <select id="my-agent-select" class="search-input" 
-                    onchange="localStorage.setItem(MY_AGENT_KEY, this.value); htmx.trigger('body', 'refresh-vals');">
-                <option value="">None</option>
-                {{range $symbol, $details := .Agents}}
-                <option value="{{$symbol}}">{{$symbol}}</option>
-                {{end}}
-            </select>
-        </div>
-    </div>
-
-    <script>
-        (function() {
-            const myAgent = localStorage.getItem('my-agent');
-            if (myAgent) {
-                const select = document.getElementById('my-agent-select');
-                if (select) {
-                    select.value = myAgent;
-                }
-            }
-        })();
-    </script>
-</div>
