--- cmd/gui/templates/stats.html
@@ -1,41 +1,0 @@
-<div class="stats-container">
-    <h2>Server Stats</h2>
-    <div class="stats-grid">
-        <div class="stat-card">
-            <h3>Status</h3>
-            <p>{{.status}}</p>
-        </div>
-        <div class="stat-card">
-            <h3>Version</h3>
-            <p>{{.version}}</p>
-        </div>
-        <div class="stat-card">
-            <h3>Reset Date</h3>
-            <p>{{.reset}}</p>
-        </div>
-        <div class="stat-card">
-            <h3>Next Reset</h3>
-            <p>{{.nextReset}}</p>
-        </div>
-        <div class="stat-card">
-            <h3>Agents</h3>
-            <p>{{.agents}}</p>
-        </div>
-        <div class="stat-card">
-            <h3>Ships</h3>
-            <p>{{.ships}}</p>
-        </div>
-        <div class="stat-card">
-            <h3>Systems</h3>
-            <p>{{.systems}}</p>
-        </div>
-        <div class="stat-card">
-            <h3>Waypoints</h3>
-            <p>{{.waypoints}}</p>
-        </div>
-        <div class="stat-card">
-            <h3>Last Market Update</h3>
-            <p>{{.marketUpdate}}</p>
-        </div>
-    </div>
-</div>
