+++ internal/collector/api.go
@@ -1,4 +1,4 @@
-package main
+package collector
 
 import (
 	"context"
