+++ internal/collector/jumpgates.go.next
@@ -1,4 +1,4 @@
-package main
+package collector
 
 import (
 	"context"
