+++ internal/collector/jumpgates.go.old
@@ -1,4 +1,4 @@
-package main
+package collector
 
 import (
 	"context"
