+++ internal/frontend/charts.go
